разделил общий файл на модули, и кажется я что то с гитхабом не правильно сделал
