+++ js/offer-data.js
@@ -1,0 +1,40 @@
+const titleArray = [
+  'Лучшее место для отдыха!',
+  'Самые красивые комнаты!',
+  'У нас бесплатные завтраки!',
+  'Самый длинный бассейн!'
+];
+
+const typeMass = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+
+const timeMass = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const featuresMass = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+
+const photoArray = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+const minLatitude = 65000;
+const maxLatitude = 70000;
+const minLongitude = 70000;
+const maxLongitude = 80000;
